[net] implementing poll for sm_behaviour, general setup works now
The general setup for sending data from peer to peer using the state
machine now works. Details of where data is kept, how to submit in the
best way still lacks a little but for now it works. A "protocol" has to be
defined that clients communicate, what is of interest.